Patch for ARM 32bit. (#1687)
Fixes some issues where 64-bit assumptions in the code cause problems on a 32-bit platform.

N.B.: There is a problem with the alignment of the sciond.connector struct in Go on ARM 32 that this does _not_ fix. https://github.com/scionproto/scion/issues/1688 is filed to track a higher-level fix for this issue.

+++ c/lib/scion/address.h
@@ -30,7 +30,7 @@ typedef struct {
 
 #define ISD_BITS 16
 #define AS_BITS 48
-#define AS_MASK ((1L << AS_BITS) - 1)
+#define AS_MASK (((uint64_t)1 << AS_BITS) - 1)
 
 #define ISD(isd_as) ((uint16_t)((isd_as) >> AS_BITS))
 #define AS(isd_as) ((isd_as) & AS_MASK)
